# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mor. HPC Cluster Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventSeverity } from './activityEventSeverity';
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  timestamp: string;
  severity: ActivityEventSeverity;
  /** @nullable */
  jobId?: number | null;
  /** @nullable */
  userId?: number | null;
}