# lib/api-zod/src/generated/types/node.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mor. HPC Cluster Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { NodeStatus } from './nodeStatus';

export interface Node {
  id: number;
  name: string;
  cpuCores: number;
  gpuCount: number;
  memoryGB: number;
  status: NodeStatus;
  cpuLoad: number;
  gpuLoad: number;
  memoryLoad: number;
  partition: string;
  updatedAt?: string;
}