# lib/api-zod/src/generated/types/resourceAvailability.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mor. HPC Cluster Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceAvailability {
  availableCpus: number;
  availableGpus: number;
  availableMemoryGB: number;
  cpuUtilPercent: number;
  gpuUtilPercent: number;
  memoryUtilPercent: number;
}